Exercise description box
Added a text area to display a description of an exercise to the user.
Made modifications to the controller to recognize click events on and off of exercises
Added visual indicator to remind users to select a difficulty
Added logic for setting and retrieving description.
Added logic to control the size of the gui and prevent user manipulation.